pyproject.toml
[project]
name = "autotrade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # 金融量化与交易
    "pyqlib",                    # 量化金融库
    "finrl",                     # 金融强化学习框架
    "lumibot",                   # 交易机器人框架
    "alpaca-py>=0.43.2",         # Alpaca 交易所 API
    "ta-lib>=0.6.8",             # 技术分析指标库
    "yfinance>=1.0",             # Yahoo Finance 数据获取
    # 机器学习框架
    "lightgbm",                  # LightGBM 梯度提升
    "catboost>=1.2.8",           # CatBoost 梯度提升
    "xgboost>=3.1.3",            # XGBoost 梯度提升
    "scikit-learn>=1.8.0",       # 机器学习工具库
    "rdagent>=0.8.0",            # 自动化量化交易研究
    # 数据处理与科学计算
    "pandas>=2.3.3",             # 数据分析库
    "pyarrow>=22.0.0",           # 列式存储格式
    "scipy>=1.17.0",             # 科学计算库
    "dask[complete]>=2026.1.1",  # 并行计算框架
    "bottleneck>=1.6.0",         # NumPy 性能加速
    # 工作流与优化
    "prefect>=3.6.11",           # 工作流编排
    "optuna>=4.6.0",             # 超参数优化
    # 可视化与Web界面
    "plotly>=6.5.2",             # 交互式图表
    "shap>=0.49.1",              # 模型可解释性
    "websockets>=12.0",          # WebSocket 通信
    "fastapi>=0.115.0",          # Web UI 框架
    "uvicorn>=0.30.0",           # ASGI 服务器
    # 因子分析
    "alphalens-reloaded>=0.4.6", # 因子绩效分析
    # 工具库
    "python-dotenv>=1.2.1",      # 环境变量管理
    "loguru>=0.7.3",             # 日志库
    "apprise>=1.9.6",            # 通知库
    "pytz>=2025.2",              # 时区处理
    "massive>=2.0.3",            # 质量数据分析
    "polars>=1.37.1",            # 高性能 DataFrame 库
    "duckdb>=1.4.3",             # 嵌入式分析数据库
    "arcticdb>=6.7.0",           # 时序数据库
    "riskfolio-lib>=7.2.0",      # 投资组合优化
    "mlflow>=3.8.1",             # 机器学习生命周期管理
    "pyod>=2.0.6",               # 异常检测
    "numba>=0.63.1",             # JIT 编译器
    "vectorbt>=0.28.2",          # 向量化回测框架
    "jinja2>=3.1.6",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

[tool.ruff]
# 行长度限制（与 Black 兼容）
line-length = 88
# 目标 Python 版本
target-version = "py311"
# 扫描的文件扩展名
include = ["*.py", "*.pyi", "**/pyproject.toml"]
# 排除目录
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# 启用的规则集
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
# 忽略的规则
ignore = [
    "E501",   # 行过长（由 formatter 处理）
    "B008",   # 函数调用作为默认参数（FastAPI 常用）
    "C901",   # 复杂度检查
]
# 允许自动修复的规则
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# 使用双引号
quote-style = "double"
# 缩进样式
indent-style = "space"
# 保留魔法尾随逗号
skip-magic-trailing-comma = false
# 自动识别行结束符
line-ending = "auto"

[tool.ruff.lint.isort]
# isort 配置
known-first-party = ["autotrade"]
force-single-line = false
lines-after-imports = 2
